+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 3.9-beta
 - Fix #241 assignments of other teachers appear in StudentAssignments.fnc.php
 - SaveTemplate() $staff_id param: use 0 for default template in Template.fnc.php
 - Add insert_attendance, update_attendance & header action hooks in Actions.php & TakeAttendance.php
+- Add $type param to AttendanceCodesTipMessage() in AttendanceCodes.fnc.php
 
 Changes in 3.8
 --------------